Created Rough Copy of Event System
Added a system by which events can be nested within an "Event_Trigger" scriptable object, which determines when an event should fire. The progression manager takes a list of all these events, and handles logistics of executing them when the time comes. The progression manager will also handle which level we are on, but I haven't quite implemented that yet.