Touch Detection Fix
Think I finally put this stupid issue of clicking objects during swipes to bed. An object can now only be tapped (with a tap handler) iff the current tap/swipe started on that object. Also tried to fix the weird issue where the name/coinname was getting overwritten in the onboarding sequence on the Android build. Haven't actually tried it yet though.